Refactored DeploymentPropertiesBuilder
- Moved SkipperDeploymentPropertiesBuilder to top class
- Updated docs to reflect changes
